+++ README.md
@@ -8,14 +8,10 @@
 The JavaScript library for generative art based on SVG.
 
 ```js
-cm.app({
-  width: 100,
-  height: 100,
-  draw: [
-    cm.svg("rect", {x: 0, y: 0, width: 100, height: 100, fill: "black"}),
-    cm.svg("circle", {cx: 50, cy: 50, r: 40, fill: "white"}),
-  ],
-}).render("#root");
+cm.create("#root", {width: 100, height: 100}).render([
+  cm.svg("rect", {x: 0, y: 0, width: 100, height: 100, fill: "black"}),
+  cm.svg("circle", {cx: 50, cy: 50, r: 40, fill: "white"}),
+]);
 ```
 
 ## Resources 📚
