+++ docs/docs/getting-started.md
@@ -48,11 +48,9 @@ In vanilla HTML, Charming can be imported as an ES module, say from jsDelivr:
 <script type="module">
   import * as cm from "https://cdn.jsdelivr.net/npm/charmingjs/+esm";
 
-  const app = cm.render({
-    // ...
-  });
+  const root = cm.create({});
 
-  document.body.append(app.node());
+  root.render(document.body, root.render([]));
 </script>
 ```
 
@@ -63,10 +61,8 @@ Charming is also available as a UMD bundle for legacy browsers.
 ```html
 <script src="https://cdn.jsdelivr.net/npm/charmingjs"></script>
 <script>
-  const app = cm.render({
-    // ...
-  });
+  const root = cm.create({});
 
-  document.body.append(app.node());
+  root.render(document.body, root.render([]));
 </script>
 ```
