+++ docs/docs/what-is-charming.md
@@ -3,20 +3,15 @@
 **Charming** (or **Charming.js**) is a free, open-source JavaScript library that creates animated and interactive SVG. Charming lets you create dynamic and expressive generative art and visualizations effortlessly. Here's a quick example that give you a sense of Charming:
 
 ```js eval t=module
-cm.app({
-  width: 200,
-  height: 50,
-  loop: true,
-  draw: () => [
-    cm.svg("circle", {
-      cx: Math.abs(Math.sin(Date.now() / 1000) * 200),
-      cy: 25,
-      r: 20,
-      stroke: "red",
-      strokeWidth: 4,
-    }),
-  ],
-}).render("#root");
+cm.create("#root", {width: 200, height: 50, loop: true}).render(() => [
+  cm.svg("circle", {
+    cx: Math.abs(Math.sin(Date.now() / 1000) * 200),
+    cy: 25,
+    r: 20,
+    stroke: "red",
+    strokeWidth: 4,
+  }),
+]);
 ```
 
 ## Based on SVG
@@ -24,14 +19,10 @@ cm.app({
 Charming provides a _svg_ function for creating SVG elements. For example, to create a white circle on a black background:
 
 ```js eval t=module
-cm.app({
-  width: 100,
-  height: 100,
-  draw: [
-    cm.svg("rect", {x: 0, y: 0, width: 100, height: 100, fill: "black"}),
-    cm.svg("circle", {cx: 50, cy: 50, r: 40, fill: "white"}),
-  ],
-}).render("#root");
+cm.create("#root", {width: 100, height: 100}).render([
+  cm.svg("rect", {x: 0, y: 0, width: 100, height: 100, fill: "black"}),
+  cm.svg("circle", {cx: 50, cy: 50, r: 40, fill: "white"}),
+]);
 ```
 
 Please refer to [Charming Mark](/docs/charming-mark) for more information.
@@ -45,26 +36,21 @@ play = Inputs.button("Replay");
 ```
 
 ```js eval t=module,replayable
-cm.app({
-  width: 100,
-  height: 100,
-  use: {transition: cm.transition},
-  draw: [
-    cm.svg("rect", {x: 0, y: 0, width: 100, height: 100, fill: "black"}),
-    cm.svg("circle", {
-      cx: 50,
-      cy: 50,
-      r: 40,
-      fill: "#4B68C9",
-      transition: {
-        keyframes: [
-          {fill: "#E5B442", r: 0, duration: 1000},
-          {fill: "#EE7A64", r: 40, duration: 2000},
-        ],
-      },
-    }),
-  ],
-}).render("#root");
+cm.create("#root", {width: 100, height: 100, use: {transition: cm.transition}}).render([
+  cm.svg("rect", {x: 0, y: 0, width: 100, height: 100, fill: "black"}),
+  cm.svg("circle", {
+    cx: 50,
+    cy: 50,
+    r: 40,
+    fill: "#4B68C9",
+    transition: {
+      keyframes: [
+        {fill: "#E5B442", r: 0, duration: 1000},
+        {fill: "#EE7A64", r: 40, duration: 2000},
+      ],
+    },
+  }),
+]);
 ```
 
 ## Incremental Updates
@@ -77,18 +63,13 @@ const height = 150;
 let x = width / 2;
 let y = height / 2;
 
-cm.app({
-  width: 600,
-  height: 150,
-  loop: true,
-  draw: () => {
-    x += cm.random(-1, 1);
-    y += cm.random(-1, 1);
-    x = cm.constrain(x, 0, width);
-    y = cm.constrain(y, 0, height);
-    return [cm.svg("circle", {cx: x, cy: y, fill: "black", r: 20})];
-  },
-}).render("#root");
+cm.create("#root", {width: 600, height: 150, loop: true}).render(() => {
+  x += cm.random(-1, 1);
+  y += cm.random(-1, 1);
+  x = cm.constrain(x, 0, width);
+  y = cm.constrain(y, 0, height);
+  return [cm.svg("circle", {cx: x, cy: y, fill: "black", r: 20})];
+});
 ```
 
 ## Reactivity for Interaction
@@ -96,21 +77,16 @@ cm.app({
 ```js eval t=module
 const state = cm.state({clicked: false});
 
-cm.app({
-  width: 100,
-  height: 100,
-  styleBackground: "black",
-  draw: () => [
-    cm.svg("circle", {
-      cx: 50,
-      cy: 50,
-      r: 40,
-      fill: state.clicked ? "red" : "white",
-      styleCursor: "pointer",
-      onClick: () => (state.clicked = !state.clicked),
-    }),
-  ],
-}).render("#root");
+cm.create("#root", {width: 100, height: 100, styleBackground: "black"}).render(() => [
+  cm.svg("circle", {
+    cx: 50,
+    cy: 50,
+    r: 40,
+    fill: state.clicked ? "red" : "white",
+    styleCursor: "pointer",
+    onClick: () => (state.clicked = !state.clicked),
+  }),
+]);
 ```
 
 ## A Collection of Tools
