+++ src/index.js
@@ -1,6 +1,6 @@
-export {app} from "./app.js";
 export {state} from "./state.js";
 export {constrain} from "./math.js";
 export {Mark, svg, html} from "./mark.js";
+export {create} from "./create.js";
 export {Renderer} from "./renderer.js";
 export * from "charmingjs-vector";
