+++ src/state.js
@@ -1,4 +1,4 @@
-import {drawRef} from "./render.js";
+import {drawRef} from "./create.js";
 
 let active;
