+++ test/apps/groupNested.js
@@ -1,16 +1,13 @@
 import * as cm from "../../src/index.js";
 
 export function groupNested() {
-  const app = cm.app({
-    draw: [
-      cm.svg("g", {
-        children: [
-          cm.svg("g", {
-            children: [cm.svg("circle", {cx: 50, cy: 50, r: 50})],
-          }),
-        ],
-      }),
-    ],
-  });
-  return app.render();
+  return cm.create().render([
+    cm.svg("g", {
+      children: [
+        cm.svg("g", {
+          children: [cm.svg("circle", {cx: 50, cy: 50, r: 50})],
+        }),
+      ],
+    }),
+  ]);
 }
